DRYed intake v1 payload decoding and cleaned up decoding generally.
