+++ model/sourcemap/payload_test.go
@@ -23,7 +23,6 @@ import (
 
 	s "github.com/go-sourcemap/sourcemap"
 	"github.com/stretchr/testify/assert"
-	"github.com/stretchr/testify/require"
 
 	"github.com/elastic/apm-server/sourcemap"
 	"github.com/elastic/apm-server/tests/loader"
@@ -36,8 +35,32 @@ func getStr(data common.MapStr, key string) string {
 	return rs.(string)
 }
 
-func TestPayloadTransform(t *testing.T) {
-	p := Payload{
+func TestDecode(t *testing.T) {
+	data, err := loader.LoadValidData("sourcemap")
+	assert.NoError(t, err)
+
+	sourcemap, err := DecodeSourcemap(data)
+	assert.NoError(t, err)
+
+	rs := sourcemap.Transform(&transform.Context{})
+	assert.Len(t, rs, 1)
+	event := rs[0]
+	assert.WithinDuration(t, time.Now(), event.Timestamp, time.Second)
+	output := event.Fields["sourcemap"].(common.MapStr)
+
+	assert.Equal(t, "js/bundle.js", getStr(output, "bundle_filepath"))
+	assert.Equal(t, "service", getStr(output, "service.name"))
+	assert.Equal(t, "1", getStr(output, "service.version"))
+	assert.Equal(t, data["sourcemap"], getStr(output, "sourcemap"))
+
+	sourcemap, err = DecodeSourcemap(nil)
+	if assert.Error(t, err) {
+		assert.Contains(t, err.Error(), "Error fetching field")
+	}
+}
+
+func TestTransform(t *testing.T) {
+	p := Sourcemap{
 		ServiceName:    "myService",
 		ServiceVersion: "1.0",
 		BundleFilepath: "/my/path",
@@ -107,29 +130,4 @@ func (a *smapMapperFake) Apply(id sourcemap.Id, lineno, colno int) (*sourcemap.M
 
 func (sm *smapMapperFake) NewSourcemapAdded(id sourcemap.Id) {
 	sm.c = map[string]*sourcemap.Mapping{}
-}
-
-func TestTransform(t *testing.T) {
-	data, err := loader.LoadValidData("sourcemap")
-	assert.NoError(t, err)
-
-	payload, err := DecodePayload(data)
-	assert.NoError(t, err)
-	require.Len(t, payload, 1)
-
-	rs := payload[0].Events(&transform.Context{})
-	assert.Len(t, rs, 1)
-	event := rs[0]
-	assert.WithinDuration(t, time.Now(), event.Timestamp, time.Second)
-	output := event.Fields["sourcemap"].(common.MapStr)
-
-	assert.Equal(t, "js/bundle.js", getStr(output, "bundle_filepath"))
-	assert.Equal(t, "service", getStr(output, "service.name"))
-	assert.Equal(t, "1", getStr(output, "service.version"))
-	assert.Equal(t, data["sourcemap"], getStr(output, "sourcemap"))
-
-	payload, err = DecodePayload(nil)
-	if assert.Error(t, err) {
-		assert.Contains(t, err.Error(), "Error fetching field")
-	}
 }
