+++ processor/metric/metric.go
@@ -24,9 +24,10 @@ import (
 )
 
 var (
-	Processor = &processor.PayloadProcessor{
-		ProcessorName: "metric",
-		DecodePayload: metric.DecodePayload,
+	Processor = &processor.EventsProcessor{
+		SingularName:  "metric",
+		PluralKeyName: "metrics",
+		EventDecoder:  metric.DecodeMetric,
 		PayloadSchema: metric.PayloadSchema(),
 		DecodingCount: monitoring.NewInt(metric.Metrics, "decoding.count"),
 		DecodingError: monitoring.NewInt(metric.Metrics, "decoding.errors"),
